fix: updated project structure a little more
